Move shared functionality from the details controller into a base controller
This is the first step in refactoring this section. There's still quite a
bit to do. The next step will separating the grade and year from into
two forms, then updating the controllers to use them.
