+++ app/controllers/candidate_interface/gcse/resolve_gcse_edit_path_concern.rb
@@ -4,7 +4,7 @@ module CandidateInterface
       def resolve_gcse_edit_path(subject)
         case subject
         when 'maths'
-          Rails.application.routes.url_helpers.candidate_interface_edit_gcse_maths_grade_path
+          candidate_interface_edit_gcse_maths_grade_path
         when 'science'
           candidate_interface_edit_gcse_science_grade_path
         when 'english'
