+++ app/controllers/candidate_interface/gcse/science/grade_controller.rb
@@ -1,10 +1,5 @@
 module CandidateInterface
-  class Gcse::Science::GradeController < CandidateInterfaceController
-    include Gcse::GradeControllerConcern
-
-    before_action :redirect_to_dashboard_if_submitted
-    before_action :set_subject
-
+  class Gcse::Science::GradeController < Gcse::BaseController
     def edit
       @gcse_grade_form = science_gcse_grade_form
       @qualification_type = gcse_science_qualification.qualification_type
