+++ app/controllers/candidate_interface/gcse/year_controller.rb
@@ -1,20 +1,21 @@
 module CandidateInterface
-  class Gcse::YearController < Gcse::DetailsController
-    before_action :redirect_to_dashboard_if_submitted
+  class Gcse::YearController < Gcse::BaseController
+    def edit
+      @application_qualification = details_form
+      @qualification_type = details_form.qualification.qualification_type
+    end
 
     def update
+      @application_qualification = details_form
       @qualification_type = details_form.qualification.qualification_type
 
       details_form.award_year = details_params[:award_year]
 
-      @application_qualification = details_form.save_year
-
-      if @application_qualification
+      if @application_qualification.save_year
         update_gcse_completed(false)
 
         redirect_to candidate_interface_gcse_review_path
       else
-        @application_qualification = details_form
         track_validation_error(@application_qualification)
 
         render :edit
